requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
graphviz==0.12
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kaggle==1.5.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.17.0
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pydot==1.4.1
Pygments==2.4.2
pyparsing==2.4.1.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-slugify==3.0.3
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.0.0
qtconsole==4.5.2
requests==2.22.0
Send2Trash==1.5.0
sip==4.19.13
six==1.12.0
tb-nightly==1.14.0a20190603
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019060501
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.0
wrapt==1.11.2
